feat: dev deploy script takes args to handle full container restart, restart from stopped, and start from clean

+++ deploy_dev.py
@@ -8,6 +8,39 @@ import asyncio
 import platform
 import qdrant_client
 from qdrant_client.models import VectorParams, Distance
+import argparse
+import docker
+
+# Get the docker client
+docker_client = docker.from_env()
+
+# ==============================================================
+# Get the args from argparse
+# ==============================================================
+
+parser = argparse.ArgumentParser()
+parser.add_argument(
+    "-rc",
+    "--restart-containers",
+    action="store_true",
+    help="Restart the docker containers (WARNING: will delete any existing containers)",
+)
+parser.add_argument(
+    "-kc",
+    "--kill-containers",
+    action="store_true",
+    help="Kill the docker containers when the script ends",
+)
+parser.add_argument(
+    "-f",
+    "--frontend",
+    action="store_true",
+    help="Start the frontend in the script",
+)
+
+args = parser.parse_args()
+
+print(args)
 
 # ==============================================================
 # Start the PostgreSQL server
@@ -15,29 +48,60 @@ from qdrant_client.models import VectorParams, Distance
 
 postgres_name = "dev-postgres"
 
-# Stop any container that exists
-subprocess.run(["docker", "stop", postgres_name])
-
-# Remove it from docker
-subprocess.run(["docker", "rm", postgres_name])
-
-# Start the PostgreSQL server
-postgres_process = subprocess.run(
-    [
-        "docker",
-        "run",
-        "--name",
-        postgres_name,
-        "-e",
-        "POSTGRES_PASSWORD=postgres",
-        "-p",
-        "5432:5432",
-        "-d",
-        "postgres",
-    ],
+# Check if the postgres container exists
+postgres_exists = (
+    len(docker_client.containers.list(filters={"name": postgres_name})) > 0
 )
 
-time.sleep(2)
+# Shutdown any pre-existing container
+if args.restart_containers:
+
+    # Stop any container that exists
+    subprocess.run(["docker", "stop", postgres_name])
+
+    # Remove it from docker
+    subprocess.run(["docker", "rm", postgres_name])
+
+
+postgres_stopped = False
+
+# Run the container
+if args.restart_containers or not postgres_exists:
+    # Start the PostgreSQL server
+    print("Starting new postgres container...")
+    postgres_process = subprocess.run(
+        [
+            "docker",
+            "run",
+            "--name",
+            postgres_name,
+            "-e",
+            "POSTGRES_PASSWORD=postgres",
+            "-p",
+            "5432:5432",
+            "-d",
+            "postgres",
+        ],
+    )
+
+    time.sleep(2)
+
+    # Check if the postgres container exists
+    postgres_exists = (
+        len(docker_client.containers.list(filters={"name": postgres_name})) > 0
+    )
+
+    # If the container doesn't exist, presumably the name is taken and it's
+    # stopped but free to run
+    if not postgres_exists:
+        # Restart the stopped container
+        print("Restarting stopped postgres container...")
+        postgres_process = subprocess.run(["docker", "start", postgres_name])
+
+        # Set postgres stopped to true
+        postgres_stopped = True
+
+        time.sleep(2)
 
 
 async def setup_postgres():
@@ -89,7 +153,13 @@ async def setup_postgres():
     print("added all tables")
 
 
-asyncio.run(setup_postgres())
+# Add the embeddings collection if they've been restarted
+if args.restart_containers:
+    asyncio.run(setup_postgres())
+
+# Add the embeddings collection if the containers have been initially created
+if not postgres_exists and not postgres_stopped:
+    asyncio.run(setup_postgres())
 
 
 # ==============================================================
@@ -98,35 +168,78 @@ asyncio.run(setup_postgres())
 
 qdrant_name = "dev-qdrant"
 
-# Stop any container that exists
-subprocess.run(["docker", "stop", qdrant_name])
-
-# Remove it from docker
-subprocess.run(["docker", "rm", qdrant_name])
-
-
-# Start the server in memory
-qdrant_process = subprocess.run(
-    [
-        "docker",
-        "run",
-        "--name",
-        qdrant_name,
-        "-p",
-        "6333:6333",
-        "-d",
-        "qdrant/qdrant",
-    ],
+# Check if the qdrant container exists
+qdrant_exists = (
+    len(docker_client.containers.list(filters={"name": "dev-qdrant"})) > 0
 )
-# qdrant = qdrant_client.QdrantClient(host="localhost", port=6333)
-# created = qdrant.create_collection(
-#     collection_name="embeddings",
-#     vectors_config=VectorParams(size=384, distance=Distance.COSINE),
-# )
 
-# print("Embeddings collection created:", created)
+# Shutdown any pre-existing container
+if args.restart_containers:
+    # Stop any container that exists
+    subprocess.run(["docker", "stop", qdrant_name])
 
-time.sleep(2)
+    # Remove it from docker
+    subprocess.run(["docker", "rm", qdrant_name])
+
+qdrant_stopeed = False
+
+# Run the container
+if args.restart_containers or not qdrant_exists:
+    # Start the server in docker
+    qdrant_process = subprocess.run(
+        [
+            "docker",
+            "run",
+            "--name",
+            qdrant_name,
+            "-p",
+            "6333:6333",
+            "-d",
+            "qdrant/qdrant",
+        ],
+    )
+
+    qdrant = qdrant_client.QdrantClient(host="localhost", port=6333)
+
+    time.sleep(2)
+
+    # Check the qdrant container exists
+    qdrant_exists = (
+        len(docker_client.containers.list(filters={"name": qdrant_name})) > 0
+    )
+
+    # If the container doesn't exist, presumably the name is taken and it's
+    # stopped but free to run
+    if not qdrant_exists:
+        # Start the stopped container
+        qdrant_process = subprocess.run(["docker", "start", qdrant_name])
+
+        # Set qdrant stopped to true
+        qdrant_stopped = True
+
+        time.sleep(2)
+
+# Add the embeddings collection if the containers have been restarted
+if args.restart_containers:
+    created = qdrant.create_collection(
+        collection_name="embeddings",
+        vectors_config=VectorParams(size=384, distance=Distance.COSINE),
+    )
+
+    print("Embeddings collection created:", created)
+
+    time.sleep(2)
+
+# Add the embeddings collection if the containers have been initially created
+if not qdrant_exists and not qdrant_stopped:
+    created = qdrant.create_collection(
+        collection_name="embeddings",
+        vectors_config=VectorParams(size=384, distance=Distance.COSINE),
+    )
+
+    print("Embeddings collection created:", created)
+
+    time.sleep(2)
 
 # ==============================================================
 # Start the FastAPI backend
@@ -147,7 +260,7 @@ fastapi_process = subprocess.Popen(
     ],
 )
 
-print(fastapi_process.pid)
+print("FastAPI backend process ID:", fastapi_process.pid)
 
 
 # ==============================================================
@@ -168,6 +281,7 @@ data["API_URL"] = f"http://localhost:{port}"
 with open("static/env.json", "w") as f:
     json.dump(data, f, indent=4)
 
+
 # ==============================================================
 # Start the frontend in dev mode
 # ==============================================================
@@ -180,13 +294,17 @@ platform_type = platform.system()
 # Get a boolean flag
 use_shell = platform_type == "Windows"
 
-# Start the frontend
-frontend_process = subprocess.Popen(
-    ["npm", "run", "dev"],
-    shell=use_shell,
-)
+frontend_process = None
 
-print(frontend_process.pid)
+# Start the frontend
+if args.frontend:
+    frontend_process = subprocess.Popen(
+        ["npm", "run", "dev"],
+        shell=use_shell,
+    )
+
+if frontend_process:
+    print("Frontend process ID:", frontend_process.pid)
 
 
 # ==============================================================
@@ -200,16 +318,18 @@ def handle_end(signal, frame):
     # Terminate the backend
     fastapi_process.terminate()
 
-    # Terminate the frontend
-    frontend_process.terminate()
+    if args.frontend:
+        # Terminate the frontend
+        frontend_process.terminate()
 
-    # Stop and remove the postgres docker container
-    subprocess.run(["docker", "stop", postgres_name])
-    subprocess.run(["docker", "rm", postgres_name])
+    if args.kill_containers:
+        # Stop and remove the postgres docker container
+        subprocess.run(["docker", "stop", postgres_name])
+        subprocess.run(["docker", "rm", postgres_name])
 
-    # Stop and remove the qdrant docker container
-    subprocess.run(["docker", "stop", qdrant_name])
-    subprocess.run(["docker", "rm", qdrant_name])
+        # Stop and remove the qdrant docker container
+        subprocess.run(["docker", "stop", qdrant_name])
+        subprocess.run(["docker", "rm", qdrant_name])
 
     # Set the .env dev to false
